Save Images in DB and retrieve them
Images can be uploaded one by one to Backend, checks if the user is the owner of the product and then saves the image to db.
To retrieve images of a product first you need to get the image ids of the product and then get the images one by one by their id.